refactor: change name some function
